Rename inference_http_endpoint -> inference_engine_client_http_endpoint
