+++ skyrl-train/tests/cpu/http/test_openai_request_utils.py
@@ -12,7 +12,7 @@ from skyrl_train.inference_engines.openai_api_protocol import (
     check_unsupported_fields,
 )
 from skyrl_train.inference_engines.base import InferenceEngineOutput
-from skyrl_train.inference_engines.inference_http_endpoint import (
+from skyrl_train.inference_engines.inference_engine_client_http_endpoint import (
     convert_openai_to_inference_input,
     convert_inference_output_to_openai,
 )
