+++ skyrl-train/tests/gpu/gpu_ci/test_inference_engine_client_http_endpoint.py
@@ -5,9 +5,9 @@ This uses the same workflow as test_policy_local_engines_e2e.py, but with the HT
 the inference client engine. Only requires 1 GPU.
 
 # Run only vllm tests (requires vllm extra):
-uv run --isolated --extra dev --extra vllm pytest tests/gpu/gpu_ci/test_inference_http_endpoint.py -m "vllm"
+uv run --isolated --extra dev --extra vllm pytest tests/gpu/gpu_ci/test_inference_engine_client_http_endpoint.py -m "vllm"
 # Run only sglang tests (requires sglang extra):
-uv run --isolated --extra dev --extra sglang pytest tests/gpu/gpu_ci/test_inference_http_endpoint.py -m "sglang"
+uv run --isolated --extra dev --extra sglang pytest tests/gpu/gpu_ci/test_inference_engine_client_http_endpoint.py -m "sglang"
 """
 
 import json
@@ -27,7 +27,7 @@ from litellm import acompletion as litellm_async_completion
 from tests.gpu.utils import init_worker_with_type, get_test_prompts
 from skyrl_train.entrypoints.main_base import config_dir
 from skyrl_train.inference_engines.utils import get_sampling_params_for_backend
-from skyrl_train.inference_engines.inference_http_endpoint import (
+from skyrl_train.inference_engines.inference_engine_client_http_endpoint import (
     serve,
     wait_for_server_ready,
     shutdown_server,
@@ -371,7 +371,7 @@ def test_http_endpoint_error_handling():
             model=MODEL,
         )
 
-        from skyrl_train.inference_engines.inference_http_endpoint import (
+        from skyrl_train.inference_engines.inference_engine_client_http_endpoint import (
             serve,
             wait_for_server_ready,
         )
